requirements.txt
ops >= 2.2.0
prometheus-client==0.16.0
Flask==2.3.2
opencv-python==4.8.0.76
picamera==1.13
Pillow==10.0.0
